+++ infrahub_sdk/node.py
@@ -25,7 +25,6 @@ if TYPE_CHECKING:
     from .schema import AttributeSchemaAPI, MainSchemaTypesAPI, RelationshipSchemaAPI
     from .types import Order
 
-# pylint: disable=too-many-lines
 
 PROPERTIES_FLAG = ["is_visible", "is_protected"]
 PROPERTIES_OBJECT = ["source", "owner"]
@@ -801,7 +800,7 @@ class InfrahubNodeBase:
         Returns:
             dict[str, Dict]: Representation of an input data in dict format
         """
-        # pylint: disable=too-many-branches
+
         data = {}
         variables = {}
 
@@ -1119,14 +1118,14 @@ class InfrahubNode(InfrahubNodeBase):
     async def artifact_generate(self, name: str) -> None:
         self._validate_artifact_support(ARTIFACT_GENERATE_FEATURE_NOT_SUPPORTED_MESSAGE)
 
-        artifact = await self._client.get(kind="CoreArtifact", definition__name__value=name, object__ids=[self.id])
+        artifact = await self._client.get(kind="CoreArtifact", name__value=name, object__ids=[self.id])
         await artifact.definition.fetch()  # type: ignore[attr-defined]
         await artifact.definition.peer.generate([artifact.id])  # type: ignore[attr-defined]
 
     async def artifact_fetch(self, name: str) -> str | dict[str, Any]:
         self._validate_artifact_support(ARTIFACT_GENERATE_FEATURE_NOT_SUPPORTED_MESSAGE)
 
-        artifact = await self._client.get(kind="CoreArtifact", definition__name__value=name, object__ids=[self.id])
+        artifact = await self._client.get(kind="CoreArtifact", name__value=name, object__ids=[self.id])
         content = await self._client.object_store.get(identifier=artifact.storage_id.value)  # type: ignore[attr-defined]
         return content
 
@@ -1251,7 +1250,6 @@ class InfrahubNode(InfrahubNodeBase):
         Returns:
             dict[str, Union[Any, Dict]]: GraphQL query in dictionary format
         """
-        # pylint: disable=too-many-branches
 
         data: dict[str, Any] = {}
 
@@ -1763,7 +1761,6 @@ class InfrahubNodeSync(InfrahubNodeBase):
         Returns:
             dict[str, Union[Any, Dict]]: GraphQL query in dictionary format
         """
-        # pylint: disable=too-many-branches
 
         data: dict[str, Any] = {}