+++ tests/integration/test_infrahub_client.py
@@ -17,10 +17,6 @@ if TYPE_CHECKING:
 
 class TestInfrahubNode(TestInfrahubDockerClient, SchemaAnimal):
     @pytest.fixture(scope="class")
-    def infrahub_version(self) -> str:
-        return "1.1.0"
-
-    @pytest.fixture(scope="class")
     async def base_dataset(
         self,
         client: InfrahubClient,