+++ tests/integration/test_infrahub_client_sync.py
@@ -24,7 +24,7 @@
 #     from infrahub.database import InfrahubDatabase
 #
 #
-# # pylint: disable=unused-argument
+#
 #
 #
 # class TestInfrahubClientSync: